requirements.txt
flask
mysql-connector-python
pdfkit